refactor: ddl 리팩토링
