+++ src/main/java/persistence/sql/ddl/domain/Table.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl;
+package persistence.sql.ddl.domain;
 
 public class Table {
